fix: do not rename error in external export
